Minor Fixes
 Update the block file counter in database when using -reindex

This problem is like earth (mostly harmless). After/during a
-reindex, it means the statistics about the last block file
reported in debug.log are always of blk00000.dat instead of the
last file. Apart from that, it means a few more database entries
need to be read when finding a file to append to the first time.


 Allow lengthy block reconnections to be interrupted

When the coin database is out of date with the block database, the
best block in it is automatically switched to. This reconnection
process can take time, so allow it to be interrupted.

This also stops block connection as soon as shutdown is requested,
leading to a faster shutdown.

BTC Commit(s):
c8b2e4426b78e63c0e9565688192963331221984
3fb9b99cca390e8ac5838c03363c46a99396e463