Make sure we always have a node to do IBD from
This introduces the concept of the 'sync node', which is the one we
asked for missing blocks. In case the sync node goes away, a new one
will be selected.

For now, the heuristic is very simple, but it can easily be extended
later to add better policies.


Add bytessent, bytesrecv and syncnode to getpeerinfo

BTC Commit(s):
6ed71b5e4faa2ea6d739eefa9c3dcc15db80439c
86648a8d16699ab392508a48bb867d3fc4f7d6cf