[QT] Don't ask for a passphrase to getnewaddress.
With an encrypted wallet the GUI was prompting for a passphrase every time
 the user requested a new address.  This is unnecessary, increases the
 exposure to keyboard sniffers, and discourages using fresh addresses for
 every transaction.

Instead only prompt for a passphrase when the keypool runs out, also call
 the new address function with the flag that prevents reuse.

Thanks to AlexNagy on IRC for pointing this out and who wouldn't take any
 lip from a curmudgeonly developer and insisted on what he knew to be true.


Remove fAllowReuse from GetKeyFromPool.

With the GUI password fix this was always false.

BTC Commit(s):
20469d83dd2fd7d8efacf94f017b926be7c92e63
71ac5052d83fcba21a09e5e2b7ad66faea6bd42a