Add verbose flag to getrawmempool
[Fee logic] Don't count txins for priority to encourage sweeping.

This changes the priority calculation to not include the size of per-txin
data including up to 110 bytes of scriptsig so that transactions which
sweep up extra UTXO don't lose priority relative to ones that don't.

I'd toyed with some other variations, but it seems like any formulation
which results in an incentive stronger than making them not count will
sometimes create incentives to add extra outputs so that you have
extra inputs to consume later.  The maximum credit is limited so that
users don't lose the disincentive to stuff random data in their
transactions, the limit of 110 is based on the size of a P2SH
redemption with a compressed public key.

This shouldn't need a staged deployment because the priority is not
used as a relay criteria, only a mining criteria.


Remove dead transaction replacement code

This dead code can be resurrected from git history if
transaction replacement is ever implemented. Keeping
dead code in the source is a bad idea, because it implies
it was tested and worked at some point, which is not true.


Refactor: move GetValueIn(tx) to tx.GetValueIn()

GetValueIn makes more sense as a CTransaction member.


Add verbose boolean to getrawmempool

Also changes mempool to store CTxMemPoolEntries
to keep track of when they enter/exit the pool.

BTC Commit(s):
d6eb259953699f5bea208ff41a0967d8ea513a70
98c7c8fd1d3712e02be0e9f2eeca7e02aa54d197
0733c1bde69c6ccfe593d2eec775d0ae32fe7140
4d707d512070ed88c888fdf625c0ae0f85f68d9b